refactor(bootloader): remove unnecessary u64 cast
- Remove redundant 'as u64' cast where type is already u64

Clippy lint: unnecessary_cast

+++ bootloader/src/main.rs
@@ -867,7 +867,7 @@ fn efi_main(handle: Handle, mut system_table: SystemTable<Boot>) -> Status {
         }
 
         (&mut *pml4_ptr)[0].set_addr(
-            PhysAddr::new(pdpt_low_frame as u64),
+            PhysAddr::new(pdpt_low_frame),
             Flags::PRESENT | Flags::WRITABLE,
         );
